catelog-service/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-cors-headers==4.3.1
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-nine==0.2.7
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.4
elastic-transport==8.10.0
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0
gunicorn==22.0.0
idna==3.4
packaging==23.2
pika==1.3.2
Pillow
psycopg2-binary
PyJWT==2.8.0
redis==5.0.7
requests==2.31.0
